refactored the code and added basic info in readme.md
